+++ tests/spl/token-extensions/programs/token-extensions/src/utils.rs
@@ -13,7 +13,6 @@ use anchor_lang::{
 };
 use anchor_spl::token_interface::spl_token_2022::{
     extension::{BaseStateWithExtensions, Extension, StateWithExtensions},
-    solana_zk_token_sdk::zk_token_proof_instruction::Pod,
     state::Mint,
 };
 use spl_tlv_account_resolution::{account::ExtraAccountMeta, state::ExtraAccountMetaList};
@@ -43,13 +42,6 @@ pub fn get_mint_extensible_extension_data<T: Extension + VariableLenPack>(
     let mint_data = account.data.borrow();
     let mint_with_extension = StateWithExtensions::<Mint>::unpack(&mint_data)?;
     let extension_data = mint_with_extension.get_variable_len_extension::<T>()?;
-    Ok(extension_data)
-}
-
-pub fn get_mint_extension_data<T: Extension + Pod>(account: &mut AccountInfo) -> Result<T> {
-    let mint_data = account.data.borrow();
-    let mint_with_extension = StateWithExtensions::<Mint>::unpack(&mint_data)?;
-    let extension_data = *mint_with_extension.get_extension::<T>()?;
     Ok(extension_data)
 }